RavenDB-16865 Prepare CreateDatabase main component
