+++ src/Raven.Studio/typescript/components/pages/resources/databases/partials/create/CreateDatabase.scss
