+++ src/Raven.Studio/typescript/components/pages/resources/databases/partials/create/CreateDatabase.tsx
@@ -1,0 +1,44 @@
+import React, { useState } from "react";
+import { Modal } from "reactstrap";
+import CreateDatabaseRegular from "./regular/CreateDatabaseRegular";
+import "./CreateDatabase.scss";
+import CreateDatabaseFromBackup from "./formBackup/CreateDatabaseFromBackup";
+
+type CreateMode = "regular" | "fromBackup";
+
+interface CreateDatabaseProps {
+    closeModal: () => void;
+}
+
+export default function CreateDatabase({ closeModal }: CreateDatabaseProps) {
+    const [createMode, setCreateMode] = useState<CreateMode>("regular");
+
+    return (
+        <Modal
+            isOpen
+            wrapClassName="bs5"
+            toggle={closeModal}
+            size="lg"
+            centered
+            autoFocus
+            fade
+            className="create-database"
+        >
+            {createMode === "regular" && (
+                <CreateDatabaseRegular
+                    closeModal={closeModal}
+                    changeCreateModeToBackup={() => setCreateMode("fromBackup")}
+                />
+            )}
+            {createMode === "fromBackup" && (
+                <CreateDatabaseFromBackup
+                    closeModal={closeModal}
+                    changeCreateModeToRegular={() => setCreateMode("regular")}
+                />
+            )}
+
+            {/* TODO remove? */}
+            <div id="PopoverContainer" className="popover-container-fix"></div>
+        </Modal>
+    );
+}
